+++ tests/objectbox-java-test/src/test/java/io/objectbox/sync/ConnectivityMonitorTest.java
@@ -1,8 +1,22 @@
+/*
+ * Copyright 2025 ObjectBox Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.objectbox.sync;
 
-import io.objectbox.sync.listener.SyncTimeListener;
 import org.junit.Test;
-
 
 import javax.annotation.Nullable;
 
@@ -11,6 +25,7 @@ import io.objectbox.sync.listener.SyncCompletedListener;
 import io.objectbox.sync.listener.SyncConnectionListener;
 import io.objectbox.sync.listener.SyncListener;
 import io.objectbox.sync.listener.SyncLoginListener;
+import io.objectbox.sync.listener.SyncTimeListener;
 
 
 import static org.junit.Assert.assertEquals;
@@ -151,6 +166,18 @@ public class ConnectivityMonitorTest {
 
         @Override
         public void setSyncTimeListener(@Nullable SyncTimeListener timeListener) {
+        }
+
+        @Override
+        public void putFilterVariable(String name, String value) {
+        }
+
+        @Override
+        public void removeFilterVariable(String name) {
+        }
+
+        @Override
+        public void removeAllFilterVariables() {
         }
 
         @Override